check_null func was written & has certain issues

+++ Password vault/password_vault_db.py
@@ -21,57 +21,72 @@ class Database:
         self.cursor.execute(create_vault)
         self.connection.commit()
 
+    def check_null(self, test):
+        if "" in test:
+            return False
+        else:
+            return True
+
 
     def new_user(self, user_id, pwd):
-        new_user_query = "INSERT INTO user VALUES(?,?)"
-        self.cursor.execute(new_user_query, (user_id, pwd))
-        self.connection.commit()
+        if check_null((user_id, pwd)):
+            new_user_query = "INSERT INTO user VALUES(?,?)"
+            self.cursor.execute(new_user_query, (user_id, pwd))
+            self.connection.commit()
 
 
     def new_entry(self, user_id, site, password):
-        new_entry_query = "INSERT INTO vault VALUES(?,?,?)"
-        self.cursor.execute(new_entry_query, (user_id, site, password))
-        self.connection.commit()
+        if check_null((user_id, site, password)):
+            new_entry_query = "INSERT INTO vault VALUES(?,?,?)"
+            self.cursor.execute(new_entry_query, (user_id, site, password))
+            self.connection.commit()
 
 
-    def remove_entry(self, user_id, del_site):
-        del_site_query = f"DELETE FROM vault WHERE user_id ='{user_id}' AND site ='{del_site}'"
-        self.cursor.execute(del_site_query)
-        self.connection.commit()
+    def remove_entry(self, user_id, del_site, del_password):
+        if check_null((user_id, del_site, del_password)):
+            del_site_query = f"DELETE FROM vault WHERE user_id ='{user_id}' AND site ='{del_site}'"
+            self.cursor.execute(del_site_query)
+            self.connection.commit()
 
 
     def remove_user(self, remove_user_id):
-        del_user_query = f"DELETE FROM user WHERE user_id ='{remove_user_id}'"
-        self.cursor.execute(del_user_query)
-        self.connection.commit()
+        if check_null((remove_user_id,)):
+            del_user_query = f"DELETE FROM user WHERE user_id ='{remove_user_id}'"
+            self.cursor.execute(del_user_query)
+            self.connection.commit()
 
 
     def update_user_passcode(self, update_user, update_pwd):
-        update_user_query = f"UPDATE user SET passcode ='{update_pwd}' WHERE user_id ='{update_user}'"
-        self.cursor.execute(update_user_query)
-        self.connection.commit()
+        if check_null((update_user, update_pwd)):
+            update_user_query = f"UPDATE user SET passcode ='{update_pwd}' WHERE user_id ='{update_user}'"
+            self.cursor.execute(update_user_query)
+            self.connection.commit()
 
 
     def update_site_password(self, update_user_id, update_site, new_password):
-        update_site_query = f"UPDATE vault set password ='{new_password}' WHERE user_id = '{update_user_id}' AND site ='{update_site}'"
-        self.cursor.execute(update_site_query)
-        self.connection.commit()
+        if check_null((update_user_id, update_site, new_password)):
+            update_site_query = f"UPDATE vault set password ='{new_password}' WHERE user_id = '{update_user_id}' AND site ='{update_site}'"
+            self.cursor.execute(update_site_query)
+            self.connection.commit()
 
 
-    def user_pass_query(self,fetch_user_id):
-        ret = f"SELECT * FROM vault WHERE user_id ='{fetch_user_id}'"
-        data = self.cursor.execute(ret)
-        return data.fetchall()
+    def user_vault_query(self,fetch_user_id):
+        if check_null((fetch_user_id,)):
+            ret = f"SELECT * FROM vault WHERE user_id ='{fetch_user_id}'"
+            data = self.cursor.execute(ret)
+            return data.fetchall()
     
 
     def user_id_query(self, fetch_user_id):
-        ret = f"SELECT * FROM user WHERE user_id ='{fetch_user_id}'"
-        data = self.cursor.execute(ret)
-        return data.fetchall()
-
+        if check_null((fetch_user_id,)):
+            ret = f"SELECT * FROM user WHERE user_id ='{fetch_user_id}'"
+            data = self.cursor.execute(ret)
+            return data.fetchall()
+    
 
     def __del__(self):
         self.connection.close()
 
 #db = Database('vault_trial.db')
 #db.new_user("george", "mypass")
+#print(db.user_vault_query("george"))
